Renamed package to com.pillar.customer
